refactor 7. GameWinningCondition 테스트코드 추가 및 LottoConfig 삭제
