--- src/main/java/lotto/configuration/LottoConfig.java
@@ -1,13 +1,0 @@
-package lotto.configuration;
-
-public class LottoConfig {
-
-	public static final int PRICE_PER_GAME = 1000;
-
-	public static final int LOTTO_MIN_NUM = 1;
-	public static final int LOTTO_MAX_NUM = 45;
-
-	public static final int NUMBER_COUNT_PER_GAME = 6;
-
-	private LottoConfig() {}
-}
