+++ src/test/java/lotto/domain/LottoGameFactoryTest.java
@@ -4,8 +4,9 @@ import org.junit.jupiter.api.Test;
 
 import java.util.stream.IntStream;
 
+import static lotto.domain.LottoGame.NUMBER_COUNT_PER_GAME;
 import static org.assertj.core.api.Assertions.assertThat;
-import static lotto.configuration.LottoConfig.NUMBER_COUNT_PER_GAME;
+
 
 class LottoGameFactoryTest {
